+++ lib/graphql/persisted_queries/compiled_queries/resolver.rb
@@ -7,6 +7,28 @@ module GraphQL
       class Resolver
         include GraphQL::PersistedQueries::ResolverHelpers
 
+        # Patch to support custom serialization
+        class GraphQL::Language::Parser # rubocop:disable Style/ClassAndModuleChildren
+          SEP = "|"
+
+          def _dump(*)
+            [
+              @graphql_str,
+              JSON.generate("filename": @filename, "max_tokens": @max_tokens, "document": @document)
+            ].join(SEP)
+          end
+
+          def self._load(args)
+            graphql_str, raw_kwargs = args.split(SEP)
+
+            new(graphql_str,
+                filename: raw_kwargs["filename"],
+                max_tokens: raw_kwargs["max_tokens"]).tap do |parser|
+              parser.instance_variable_set(:@document, raw_kwargs["document"])
+            end
+          end
+        end
+
         def initialize(schema, extensions)
           @schema = schema
           @extensions = extensions